clean-locations-main/clean-locations-main/requirements.txt
boto3==1.34.109
botocore==1.34.109
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
jmespath==1.0.1
numpy==1.26.4
pandas==2.2.2
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
python-dotenv==1.0.1
langchain-anthropic==0.1.23
langchain-core==0.2.38